+++ src/main/java/ui/scenario/action/EditActionViewController.java
@@ -277,8 +277,8 @@ public class EditActionViewController implements NewSignalViewController.OnNewSi
      *
      * @return duration value or 0
      */
-    private int getCurrentEffectDuration() {
-        return inputEffectTime != null && inputEffectTime.getLength() > 0 ? Integer.valueOf(inputEffectTime.getText()) : 0;
+    private float getCurrentEffectDuration() {
+        return inputEffectTime != null && inputEffectTime.getLength() > 0 ? Float.valueOf(inputEffectTime.getText()) : 0;
     }
 
     /**
@@ -296,8 +296,8 @@ public class EditActionViewController implements NewSignalViewController.OnNewSi
      *
      * @return duration value or 0
      */
-    private int getScoreLostValue() {
-        return inputLostValue != null && inputLostValue.getLength() > 0 ? Integer.valueOf(inputLostValue.getText()) : 0;
+    private float getScoreLostValue() {
+        return inputLostValue != null && inputLostValue.getLength() > 0 ? Float.valueOf(inputLostValue.getText()) : 0;
     }
 
     /**
